Tarefa1_2022li1g013

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 2022/23.

Função Principal - mapaValido

mapaValido

Arguments

:: Mapa
argumento assume-se do tipo Mapa

-> Bool
resultado, Verdadeiro ou Falso

A função mapaValido verifica se o Mapa é válido (usa as funções auxiliares mencionadas abaixo)

A função pode ser definida da seguinte forma :

mapaValido l(Mapa larg ((Rio v , obstaculos):t)) = obstaculosinvalidos l == True && rioscontiguos l == True && larguraobstaculos l == True && minobstaculos  l == True && largmapa l == True && maxriosestradas l == True && mapaValido (Mapa larg t) 
mapaValido l(Mapa larg ((Estrada v , obstaculos):t)) = obstaculosinvalidos l == True && larguraobstaculos l == True && minobstaculos  l == True && largmapa l == True && maxriosestradas l == True && mapaValido (Mapa larg t) 
mapaValido l@(Mapa larg ((Relva , obstaculos):t)) = obstaculosinvalidos l == True &&  minobstaculos  l == True && largmapa l == True && maxriosestradas l == True && mapaValido (Mapa larg t) 

Funções auxiliares para mapaValido

obstaculosinvalidos

obstaculosinvalidos

Arguments

:: Mapa
argumento assume-se do tipo Mapa

-> Bool
resultado, Verdadeiro ou Falso

A função obstaculosinvalidos verifica se não existem obstáculos em terrenos impróprios do Mapa

A função pode ser definida da seguinte forma :

obstaculosinvalidos (Mapa larg []) = True 
obstaculosinvalidos (Mapa larg ((Rio v , obstaculos):t)) 
    | v == 0 = False 
    | Carro elem obstaculos|| Arvore elem obstaculos = False 
    | otherwise = obstaculosinvalidos (Mapa larg t) 
obstaculosinvalidos (Mapa larg ((Estrada v , obstaculos):t))  
    | v == 0 = False 
    | Tronco elem obstaculos|| Arvore elem obstaculos = False 
    | otherwise = obstaculosinvalidos (Mapa larg t) 
obstaculosinvalidos (Mapa larg ((Relva , obstaculos):t)) 
    | Carro elem obstaculos|| Tronco elem obstaculos = False 
    | otherwise = obstaculosinvalidos (Mapa larg t)

rioscontiguos

rioscontiguos

Arguments

:: Mapa
argumento assume-se do tipo Mapa

-> Bool
resultado, Verdadeiro ou Falso

A função rioscontiguos verifica se rioscontíguos têm direções opostas no Mapa

A função pode ser definida da seguinte forma :

rioscontiguos (Mapa larg ((Rio v , obstaculos):(Rio v1 , obstaculos1):t))
    | v>0 && v1>0 || v<0 && v1<0 = False
    | otherwise = rioscontiguos (Mapa larg t)
rioscontiguos (Mapa larg ((Rio _,_ ):t)) = True

larguraobstaculos

larguraobstaculos

Arguments

:: Mapa
argumento assume-se do tipo Mapa

-> Bool
resultado, Verdadeiro ou Falso

A função larguraobstaculos verifica se os troncos têm, no máximo, 5 unidades de comprimento e se os carros têm, no máximo, 3 unidades de comprimento no Mapa

A função pode ser definida da seguinte forma :

larguraobstaculos (Mapa larg ((Rio v, obstaculos):t)) = contagemTroncos obstaculos 0 
larguraobstaculos (Mapa larg ((Estrada v, obstaculos):t)) = contagemCarros obstaculos 0 

funções auxiliares da função larguraobstaculos ---> contagemTroncos e contagemCarros

contagemTroncos

Arguments

:: [Obstaculo]
pede uma lista de obstáculos

-> Int
pede um número inteiro

-> Bool
devolve Verdade ou Falso

A função contagemTroncos verifica se os troncos têm, no máximo, 5 unidades de comprimento no Mapa

A função pode ser definida da seguinte forma :

contagemTroncos :: [Obstaculo] -> Int -> Bool
contagemTroncos l 6 = False
contagemTroncos [] n = True
contagemTroncos (h:t) n
            | h == Tronco = contagemTroncos t (n+1)
            | otherwise = contagemTroncos t 0

contagemCarros

Arguments

:: [Obstaculo]
pede uma lista de obstáculos

-> Int
pede um número inteiro

-> Bool
devolve Verdade ou Falso

A função contagemCarros verifica se os carros têm, no máximo, 3 unidades de comprimento no Mapa

A função pode ser definida da seguinte forma :

contagemCarros :: [Obstaculo] -> Int -> Bool
contagemCarros l 4 = False
contagemCarros [] n = True
contagemCarros (h:t) n
            | h == Tronco = contagemCarros t (n+1)
            | otherwise = contagemCarros t 0

minobstaculos

minobstaculos

Arguments

:: Mapa
argumento assume-se do tipo Mapa

-> Bool
resultado, Verdadeiro ou Falso

A função minobstaculos verifica se existe pelo menos um obstáculo Nenhum em cada linha do Mapa

A função pode ser definida da seguinte forma :

minobstaculos (Mapa larg []) = True  
minobstaculos (Mapa larg ((Rio v , obstaculos):t))  = Nenhum elem obstaculos && minobstaculos (Mapa larg t) 
minobstaculos (Mapa larg ((Estrada v , obstaculos):t))  = Nenhum elem obstaculos && minobstaculos (Mapa larg t) 
minobstaculos (Mapa larg ((Relva , obstaculos):t))  = Nenhum elem obstaculos && minobstaculos (Mapa larg t) 

largmapa

largmapa

Arguments

:: Mapa
argumento assume-se do tipo Mapa

-> Bool
resultado, Verdadeiro ou Falso

A função largmapa verifica se o comprimento da lista de obstáculos de cada linha corresponde exatamente à largura do Mapa

A função pode ser definida da seguinte forma :

largmapa (Mapa larg []) = True 
largmapa (Mapa larg ((Rio v , obstaculos):t)) = larg == length obstaculos && largmapa (Mapa larg t) 
largmapa (Mapa larg ((Estrada v , obstaculos):t)) = larg == length obstaculos && largmapa (Mapa larg t) 
largmapa (Mapa larg ((Relva , obstaculos):t)) = larg == length obstaculos && largmapa (Mapa larg t) 

maxriosestradas

maxriosestradas

Arguments

:: Mapa
argumento assume-se do tipo Mapa

-> Bool
resultado, Verdadeiro ou Falso

A função maxriosestradas verifica se não existem, contiguamente, mais do que 4 rios, 5 estradas ou 5 relvas no Mapa

A função pode ser definida da seguinte forma :

maxriosestradas (Mapa larg ((Rio _ , _):(Rio _ , _):(Rio _ , _):(Rio _ , _):(Rio _ , _):t)) = False 
maxriosestradas (Mapa larg ((Estrada _ , _):(Estrada _ , _):(Estrada _ , _):(Estrada _ , _):(Estrada _ , _):(Estrada _ , _):t)) = False 
maxriosestradas (Mapa larg ((Relva , _):(Relva , _):(Relva , _):(Relva , _):(Relva , _):(Relva , _):t)) = False 
maxriosestradas (Mapa larg []) = True 
maxriosestradas (Mapa larg ((Rio _, _ ):t)) = True 
maxriosestradas (Mapa larg ((Estrada _, _ ):t)) = True 
maxriosestradas (Mapa larg ((Relva, _ ):t)) = True