Tarefa2_2022li1g013

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 2022/23.

Função Principal - estendeMapa

estendeMapa

Arguments

:: Mapa
recebe o Mapa

-> Int
recebe um número aleatório

-> Mapa
devolve o mapa com a nova linha adicionada

A função estendeMapa adiciona uma nova linha ao 'Mapa? de forma aleatória

A função pode ser definida da seguinte forma :

estendeMapa (Mapa larg terr) n = Mapa larg ((randomvelocidade n terreno t, listarobstaculos n larg terreno): terr)
    where terreno = escolheTerreno n (proximosTerrenosValidos (Mapa larg terr))
          t | null terr = Relva
            | otherwise = fst(head terr)

estendeMapa'

Arguments

:: Int
recebe a largura do Mapa

-> [(Terreno, [Obstaculo])]
recebe uma lista de tuplos Terreno e lista de Obstáculos

-> Int
recebe um número aleatório

-> [(Terreno, [Obstaculo])]
devolve a lista de tuplos Terreno e lista de Obstáculos com a nova linha adicionada

A função estandeMapa' é semelhante à função estendeMapa , mas em vez de receber um Mapa recebe uma lista de tuplos [(Terreno, [Obstaculo])] (usada na Tarefa 6)

A função pode ser definida da seguinte forma :

estendeMapa' larg terr n = ((randomvelocidade n terreno t, listarobstaculos n larg terreno): terr)
    where terreno = escolheTerreno n (proximosTerrenosValidos (Mapa larg terr))
          t | null terr = Relva
            | otherwise = fst(head terr)

proximosTerrenosValidos

Arguments

:: Mapa
recebe o Mapa

-> [Terreno]
devolve a lista de Terrenos válidos para a próxima linha

A função proximosTerrenosValidos serve para listas os Terrenos válidos para a próxima linha do Mapa

A função pode ser definida da seguinte forma :

proximosTerrenosValidos :: Mapa -- ^ recebe o Mapa
                        -> [Terreno] -- ^ devolve a lista de Terrenos válidos para a próxima linha
proximosTerrenosValidos (Mapa _ ((Estrada _, _):(Estrada _, _):(Estrada _, _):(Estrada _, _):(Estrada _,_):t)) = [Relva, Rio 0]
proximosTerrenosValidos (Mapa _ ((Relva,_):(Relva,_):(Relva,_):(Relva,_):(Relva,_):t)) = [Rio 0, Estrada 0]
proximosTerrenosValidos (Mapa _ ((Rio _, _):(Rio _, _):(Rio _, _):(Rio _, _):t)) = [Estrada 0, Relva]
proximosTerrenosValidos (Mapa _ _) = [Relva, Rio 0, Estrada 0]

proximosTerrenosValidos'

Arguments

:: [(Terreno, [Obstaculo])]

- ^ recebe uma lista de tuplos Terreno e lista de Obstáculos

-> [Terreno]
devolve a lista de Terrenos válidos para a próxima linha

A função proximosTerrenosValidos' é semelhante à função proximosTerrenosValidos, mas em vez de receber um Mapa recebe uma lista de tuplos [(Terreno, [Obstaculo])] (usada na Tarefa 6)

A função pode ser definida da seguinte forma :

proximosTerrenosValidos' :: [(Terreno,[Obstaculo])] 
                         -> [Terreno]
proximosTerrenosValidos' ((Estrada _, _):(Estrada _, _):(Estrada _, _):(Estrada _, _):(Estrada _,_):t) = [Relva, Rio 0]
proximosTerrenosValidos' ((Relva,_):(Relva,_):(Relva,_):(Relva,_):(Relva,_):t) = [Rio 0, Estrada 0]
proximosTerrenosValidos' ((Rio _, _):(Rio _, _):(Rio _, _):(Rio _, _):t) = [Estrada 0, Relva]
proximosTerrenosValidos' _ = [Relva, Rio 0, Estrada 0]

escolheTerreno

Arguments

:: Int
recebe um número aleatório

-> [Terreno]
recebe a lista de Terrenos válidos para a próxima linha

-> Terreno
devolve o Terreno da prócima linha do Mapa

A função escolheTerreno serve para escolher de forma ateatória o Terreno da próxima linha do Mapa

A função pode ser definida da seguinte forma :

escolheTerreno :: Int 
               -> [Terreno] 
               -> Terreno
escolheTerreno n (h:t) | mod n 4 == 0 = h
                       | mod n 4 == 1 = head t
                       | otherwise = last t

randomvelocidade

Arguments

:: Int
recebe um número aleatório

-> Terreno
recebe o Terreno da próxima linha do Mapa

-> Terreno
recebe o Terreno da primeira linha do Mapa atual

-> Terreno
devolve o Terreno e a velocidade da próxima linha do Mapa

A função randomvelocidade serve para escolher de forma ateatória a velocidade do Terreno da próxima linha do Mapa

A função pode ser definida da seguinte forma :

randomvelocidade :: Int 
                 -> Terreno 
                -> Terreno 
                -> Terreno
randomvelocidade _ Relva _ = Relva
randomvelocidade n (Rio v) (Rio v1) 
    | v1 > 0 = if even n then Rio (-2) else Rio (-1)
    | otherwise =  if even n then Rio 1 else Rio 2
randomvelocidade n (Rio v) _ 
    | mod n 5 == 0 = Rio (-1)
    | mod n 5 == 1 = Rio 1
    | mod n 5 == 2 = Rio (-2)
    | otherwise = Rio 2
randomvelocidade n (Estrada v) _ 
    | mod n 5 == 0 = Estrada (-1)
    | mod n 5 == 1 = Estrada 1
    | mod n 5 == 2 = Estrada (-2)
    | otherwise = Estrada 2

proximosObstaculosValidos

Arguments

:: Int
recebe a largura do Mapa

-> (Terreno, [Obstaculo])
recebe o tuplo Terreno e lista de Obstáculos da nova linha do Mapa

-> [Obstaculo]
devolve a lista de obstáculos que se podem adicionar a nova linha do Mapa

A função proximosObstaculosValidos serve para listar os obstáculos válidos para a próxima linha do Mapa

A função pode ser definida da seguinte forma :

proximosObstaculosValidos :: Int 
                          -> (Terreno, [Obstaculo]) 
                          -> [Obstaculo]
proximosObstaculosValidos larg (Rio _, []) = [Tronco,Nenhum]
proximosObstaculosValidos larg (Estrada _, []) = [Carro,Nenhum]
proximosObstaculosValidos larg (Relva , obs) | length obs >= larg = [] 
                                             | otherwise = if elem Nenhum obs then [Arvore, Nenhum] else [Nenhum]
proximosObstaculosValidos larg (Rio _, obs) | length obs >= larg = []
                                            | length obs > 4 = if elem Nenhum (take 5 obs) then [Tronco,Nenhum] else [Nenhum]
                                            | otherwise = if elem Nenhum obs && not(elem Tronco obs) then [Tronco] else if elem Nenhum obs then [Tronco,Nenhum] else [Nenhum] 
proximosObstaculosValidos larg (Estrada  _, obs) | length obs >= larg = []
                                                 | length obs > 2 = if elem Nenhum (take 3 obs) then [Carro,Nenhum] else [Nenhum]
                                                 | otherwise = if elem Nenhum obs then [Carro, Nenhum] else [Nenhum]

escolherobstaculo

Arguments

:: Int
recebe um número aleatório

-> Int
recebe a largura do Mapa

-> (Terreno, [Obstaculo])
recebe o tuplo Terreno e lista de Obstáculos da nova linha do Mapa

-> Obstaculo
devolve o obstáculo escolhida para se adidionar a nova linha do Mapa

A função escolherobstaculo serve para escolher um obstáculo de forma aleatória daqueles listados a partir da função proximosObstaculosValidos

A função pode ser definida da seguinte forma :

escolherobstaculo :: Int 
                  -> Int 
                  -> (Terreno, [Obstaculo]) 
                  -> Obstaculo
escolherobstaculo n larg (t,obs) | mod n 3 == 0 && mod n 2 == 0 = last obsval
                                 | mod n 3 == 0 = head obsval
                                 | otherwise = last obsval
    where obsval = proximosObstaculosValidos larg (t,obs)

listarobstaculos

Arguments

:: Int
recebe um número aleatório

-> Int
recebe a largura do Mapa

-> Terreno
recebe o Terreno da nova linha do Mapa

-> [Obstaculo]
devolve a lista de obstáculos da nova linha do Mapa

A função listarobstaculos serve para listar os obstáculos da próxima linha do Mapa

A função pode ser definida da seguinte forma :

listarobstaculos :: Int 
                 -> Int 
                 -> Terreno 
                 -> [Obstaculo]
listarobstaculos n 0 t = []
listarobstaculos n larg t = escolherobstaculo n larg (t,obs) : obs
    where obs = listarobstaculos (n + (mod n 5)) (larg-1) t