Tarefa4_2022li1g013

Description

Módulo para a realização da Tarefa 4 do projeto de LI1 em 2022/23.

Documentation

jogoTerminou

Arguments

:: Jogo
recebe o jogo

-> Bool
devolve um boolean

A função jogoTerminou é usada para verificar se o Jogo terminou devido ao Jogador estar fora do Mapa, ter caído na Água ou estar debaixo de um Carro

A função pode ser definida da seguinte forma :

jogoTerminou (Jogo (Jogador (x,y)) (Mapa l terr)) 
    | x || x= l || y || y= length terr = True 
    | otherwise = caiunaagua (x,y) (terr !! y) || debaixodecarro (x,y) (terr !! y) 

caiunaagua

Arguments

:: Coordenadas
recebe as coordenadas do jogador

-> (Terreno, [Obstaculo])
recebe um tuplo de terreno e lista de obstáculos

-> Bool
devolve um boolean

A função caiunaagua é usada para verificar se o Jogador caiu na Água

A função pode ser definida da seguinte forma :

caiunaagua (x,y) (Rio v, obstaculos)
    | obstaculos !! x == Nenhum = True
    | otherwise = False
caiunaagua _ _ = False

debaixodecarro

Arguments

:: Coordenadas
recebe as coordenadas do jogador

-> (Terreno, [Obstaculo])
recebe um tuplo de terreno e lista de obstáculos

-> Bool
devolve um boolean

A função debaixodecarro é usada para verificar se o Jogo terminou devido ao Jogador estar debaixo de um Carro

A função pode ser definida da seguinte forma :

debaixodecarro (x,y) (Estrada v, obstaculos) 
    | obstaculos !! x == Carro = True
    | otherwise = False
debaixodecarro _ _ = False